+++ jaxley/synapses/__init__.py
@@ -1,18 +1,23 @@
 # This file is part of Jaxley, a differentiable neuroscience simulator. Jaxley is
 # licensed under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
+from jaxley.synapses.conductance import ConductanceSynapse
+from jaxley.synapses.current import CurrentSynapse
+from jaxley.synapses.dynamic import DynamicSynapse
 from jaxley.synapses.ionotropic import IonotropicSynapse
 from jaxley.synapses.spike import SpikeSynapse
 from jaxley.synapses.synapse import Synapse
-from jaxley.synapses.tanh_conductance import TanhConductanceSynapse
-from jaxley.synapses.tanh_rate import TanhRateSynapse
 from jaxley.synapses.test import TestSynapse
 
 __all__ = [
     "IonotropicSynapse",
     "SpikeSynapse",
     "Synapse",
+    "DynamicSynapse",
+    "Synapse",
     "TanhConductanceSynapse",
     "TanhRateSynapse",
+    "ConductanceSynapse",
+    "CurrentSynapse",
     "TestSynapse",
 ]