+++ tests/test_syn.py
@@ -8,10 +8,20 @@ jax.config.update("jax_platform_name", "cpu")
 
 
 import numpy as np
+import pytest
+from jax.nn import relu
 
 import jaxley as jx
+from jaxley.channels import HH, Leak
 from jaxley.connect import connect
-from jaxley.synapses import IonotropicSynapse
+from jaxley.synapses import (
+    ConductanceSynapse,
+    CurrentSynapse,
+    DynamicSynapse,
+    IonotropicSynapse,
+    Synapse,
+    TestSynapse,
+)
 
 
 def test_set_and_querying_params_one_type(SimpleNet):
@@ -63,3 +73,47 @@ def test_spike_synapse(SpikeNet):
     recordings = np.asarray(jx.integrate(net, delta_t=dt))
     assert np.invert(np.any(np.isnan(recordings)))
     assert np.invert(np.all(recordings[0] == recordings[0][0]))
+
+
+@pytest.mark.parametrize(
+    "synapse", [CurrentSynapse, ConductanceSynapse, DynamicSynapse, IonotropicSynapse]
+)
+def test_synapse_correctness(SimpleNet, synapse):
+    net = SimpleNet(2, 1, 1)
+    connect(net.cell(0), net.cell(1), synapse())
+    net.insert(Leak())
+    net.cell(1).record()
+
+    i_delay = 5.0
+    i_dur = 10.0
+    i_amp = 0.1
+    net.cell(0).stimulate(jx.step_current(i_delay, i_dur, i_amp, 0.025, 20.0))
+
+    v = jx.integrate(net)
+    # Test whether the voltage increases after the synaptic current reaches the
+    # post-synaptic site.
+    assert v[0, int(i_delay * 40)] < v[0, int((i_delay + i_dur) * 40)]
+
+
+@pytest.mark.parametrize(
+    "synapse", [CurrentSynapse, ConductanceSynapse, DynamicSynapse]
+)
+@pytest.mark.parametrize(
+    "nonlinearity",
+    [
+        relu,
+        lambda x: x**2,
+    ],
+)
+def test_synapse_nonlinearity(SimpleNet, synapse, nonlinearity):
+    net = SimpleNet(2, 1, 1)
+
+    connect(net.cell(0), net.cell(1), synapse(nonlinearity))
+    net.insert(Leak())
+    net.cell(1).record()
+    net.cell(0).stimulate(jx.step_current(5.0, 10.0, 0.1, 0.025, 20.0))
+
+    v = jx.integrate(net)
+
+    # Voltage should increase after synaptic current arrives
+    assert v[0, int(5.0 * 40)] < v[0, int(15.0 * 40)]